// src/chu-de/interfaces/chuDe.interface.ts
export interface IChuDe {
  readonly id?: number;
  readonly idSyllabus?: number;
  readonly idLKHGD?: number;
  readonly ma?: string;
  readonly ten?: string;
  readonly tuan?: number;
  readonly createdAt?: Date;
  readonly updatedAt?: Date;
  readonly createdBy?: number;
  readonly updatedBy?: number;
  readonly isDeleted?: boolean;
}